Pull `has_sources` logic up into `Target`.
